refactor: broke up config into state & config
Introduced the ApplicationState struct to hold the config, update
manager and ui handler. I created this instead of storing the update
manager and ui handler on the config. This way the config is only
configuration values and the state holds the config and instances the
application needs to run. And we pass around the ApplicationState
instead of the config.