feat: adding http client
Added a local client module that is a wrapper around http.Client that
uses `github.com/dcos/dcos-go/dcos/http/transport` to set the
`Authorization` header for all requests by using a IAM config and cert
from the cluster. I used `dcos-backup` as the template for this module.
I then updated all http calls to use this wrapper and removed the
apiToken code I had added to test with.