refactor: poll zk instead of using watches
We were informed that zk watches are unreliable and you should always
also poll for changes if you are using them. Given that we decided to
switch to a polling only solution. This implementation is simpler with
less error cases.

initCurrentVersion() was updated to create the node if it doesn't
already exist. This allows polling to always do a get and not worry
about doing Exists checks. It also allows the UpdateCurrentVersion()
function to assume the node exists and always perform a Set instead of
checking if the node exists first.

initCurrentVersion() was also updated to panic if there are errors
interacting with zookeeper, if Exists(), Get() or Create() calls return
an error the service will panic. The assumption being if we can't read
or write to zookeeper we can't perform updates.