refactor(zookeeper): Use logrus
Updated zookeeper client and version store to both use `logrus` for
logging.

Added additional directory and file names to info log messages for
downloader.Client to try and make the messages more relevant. We may
still want to move these to `Debug` if they are too noisy, but that
could be useful in production if we have issues with an update.