refactor: better HTTP codes on update failure
Updated the updateManager and downloader clients to use predefined
errors for failure cases. Since I'm returning these errors instead of
wrapping the original error, I added new Error log statements so that
the original cause of the error is still captured, but we don't leak
this information through the API.

I also updated tests to use the test_helper when I was making other edits to
the tests.

- Closes DCOS-45354