fix: removing docker
We had zookeeper tests that used docker to start a zk server, but these
could only be run locally and not in CI. Then today the docker
dependency broke building due to a conflict with logrus. Since we don't
actually run this test in CI I'm remove it and all the helpers to fix
the build.