feat: use symlinks instead of serving ui assets
Changed the update process to update a symlink that is used by nginx to
server the UI. This service will no longer server the ui assets
directly. To support this new config values were added for the symlink
and a temporary symlink used during updating.

The uiFileHandler was removed since it is no longer needed.

nginx was added to the docker-compose and can be used to test calling
the ui-service and confirming the version was updated correctly.

Unit tests were updated to account for all of these changes. Since afero
doesn't support symlinks in most cases we have to use the OS file system
now, which requires the tests not be run in parallel. The test use a
sandbox folder in `./testdata` and the `tearDown` function is defered to
remove each test's sandbox folder.

`GET /api/v1/version/` endpoint was added. This was added mostly to give
an easy endpoint to call for tests, but it should also be useful in
other cases.

Several files/folders were removed that are no longer required.

Closes DCOS-46389