feat: try to create VersionsRoot on startup
Attempt to create the VersionsRoot directory on start-up if it does not
exist. If this failed we log an error, but I considered panicking. I
didn't because that would just cause the service to continually crash on
startup. We will want to make the service unhealthy if the VersionsRoot
doesnt exist when we add health checks.