feat(zookeeper): Create Node Watchers
Created abstractions for watching a zookeeper value or parent node. The
abstractions are used with a listener that is called when ever the
provided node is updated. It create a zk watch, but also polls the node
for changes on a timeout since zk watches are AT MOST once delivery.

A watcher can be long-lived, like the current use case where zkVersion
creates a watch for the version node and never closes it. Or it could be
created and closed when no longer needed.

CLOSES DCOS-46390