refactor(dcos): use dcos-go
Update the local DCOS package to use `dcos-go` package and expose a
MesoID & DetectIP functions. This will be used for reporting update status in zk to
coordinate between nodes for updating.

Renamed `localCosmos` to `clusterProxy` as its now also used to get the
mesos state locally.

I made several changes to rerun and docker-compose.yml here to help with
running locally and utilizing the `dcos-go` package.