feat: synchronized reset
Replacing local reset and update of ZK `/version` node with a
synchronized reset process. The reset is lead by the master that
receives the API call. Then the leader updates ZK `/cluster-status` node
with a status value for the cluster state machine. Each master watches
this node and when the state changes from `Idle` it starts the requested
operation (Reset or eventually Update). They then report their "node"
status of the operation state machine. The cluster then watches for
changes in node status to synchronize and transition the cluster status
based on the node statuses.

This implementation is function for the `/v1/reset/` endpoint, but is
lacking unit tests. It is also lacking some clean-up checks on start-up
to handle edge cases where ZK or any master crash during an update.